Initial proto-sync of experimental relay.
This code *does not work* in this environment. However this presents us with a first opportunity to take experimental relay for a ride in the context of Github to determine what will be necessary to get experimental relay operational.

Do not expect this code to run. Do not expect commits on this branch to be merged into master. This is purely exploratory work.